上传使用说明

+++ src/main/java/com/xp/agent/util/Maths.java
@@ -1,7 +1,7 @@
 package com.xp.agent.util;
 
 public class Maths {
-	public static long BToM(long num){
-		return num/1024/1024;
+	public static double BToM(long num){
+		return num/1024/1024.0;
 	}
 }
